feat: THARSIS-4 assigned pages folder and integrated a Chatbot Test with DialowFlow
