+++ client/package.json
@@ -18,7 +18,8 @@
     "flowbite": "^1.5.4",
     "flowbite-react": "^0.3.4",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.4.3"
   },
   "devDependencies": {
     "@types/react": "^18.0.24",
