+++ client/src/pages/Dashboard.tsx
@@ -1,0 +1,12 @@
+import ChatBot from "../components/ChatBot";
+
+const Dashboard = ()=>{
+    return(
+        <>
+            Dashboard 
+            <ChatBot/>
+        </>
+    )
+}
+
+export default Dashboard;
