+++ client/src/pages/NotFound.tsx
@@ -1,0 +1,9 @@
+const NotFound = ()=>{
+    return(
+        <>
+            404
+        </>
+    )
+}
+
+export default NotFound;
